requirements.txt
Flask
nltk
Pillow
easyocr
numpy
